Drop textstat dependency; determine speed based on starting time
Drop the dependency on textstat because it's potentially hard to
package into the plugin, and it's also not necessarily better than using
the traditional definition of wpm (five characters per word) because the
traditional definition is the one most people are familiar with, and we
don't want to mislead them, even if it turns out that it's an inferior
measure.

Determine the speed of the user based on the time that they most
recently started typing, rather than using fixed-sized windows. That
is, if the user started typing five seconds ago, only calculate the wpm
for the last five seconds, rather than a number like ten seconds or a
minute.