Add strokes meter
This adds a meter that tells you how many strokes, on average, you take
to write words. If you write out many words, this number will be higher;
likewise, if you write words in relatively few strokes, this number will
be lower.

This commit also adds a new method of counting words, which is to count
the number of space-delimited substrings.